+++ src/data.rs
@@ -8,19 +8,18 @@
 
 //! Types describing various forms of payload data.
 
-use bytes::BytesMut;
-use std::convert::TryFrom;
+use std::{convert::TryFrom, fmt};
 
 /// Data received from the remote end.
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum Incoming {
+pub enum Incoming<'a> {
     /// Text or binary data.
     Data(Data),
     /// Data sent with a PONG control frame.
-    Pong(Data)
+    Pong(&'a [u8])
 }
 
-impl Incoming {
+impl Incoming<'_> {
     /// Is this text or binary data?
     pub fn is_data(&self) -> bool {
         if let Incoming::Data(_) = self { true } else { false }
@@ -48,97 +47,59 @@ impl Incoming {
             false
         }
     }
-}
 
-impl AsRef<[u8]> for Incoming {
-    fn as_ref(&self) -> &[u8] {
+    /// The length of data (number of bytes).
+    pub fn len(&self) -> usize {
         match self {
-            Incoming::Data(d) => d.as_ref(),
-            Incoming::Pong(d) => d.as_ref()
+            Incoming::Data(d) => d.len(),
+            Incoming::Pong(d) => d.len()
         }
     }
 }
-
-impl AsMut<[u8]> for Incoming {
-    fn as_mut(&mut self) -> &mut [u8] {
-        match self {
-            Incoming::Data(d) => d.as_mut(),
-            Incoming::Pong(d) => d.as_mut()
-        }
-    }
-}
-
-impl Into<Data> for Incoming {
-    fn into(self: Incoming) -> Data {
-        match self {
-            Incoming::Data(d) => d,
-            Incoming::Pong(d) => d
-        }
-    }
-}
-
-/// Application data.
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct Data(DataRepr);
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-enum DataRepr {
-    /// Binary data.
-    Binary(BytesMut),
-    /// UTF-8 encoded data.
-    Text(BytesMut)
+pub enum Data {
+    /// Textual data (number of bytes).
+    Text(usize),
+    /// Binary data (number of bytes).
+    Binary(usize)
 }
 
 impl Data {
-    /// Create a new binary `Data` value.
-    pub(crate) fn binary(b: BytesMut) -> Self {
-        Data(DataRepr::Binary(b))
-    }
-
-    /// Create a new textual `Data` value.
-    pub(crate) fn text(b: BytesMut) -> Self {
-        Data(DataRepr::Text(b))
+    /// Is this text data?
+    pub fn is_text(&self) -> bool {
+        if let Data::Text(_) = self { true } else { false }
     }
 
     /// Is this binary data?
     pub fn is_binary(&self) -> bool {
-        if let DataRepr::Binary(_) = self.0 { true } else { false }
+        if let Data::Binary(_) = self { true } else { false }
     }
 
-    /// Is this UTF-8 encoded textual data?
-    pub fn is_text(&self) -> bool {
-        if let DataRepr::Text(_) = self.0 { true } else { false }
-    }
-}
-
-impl AsRef<[u8]> for Data {
-    fn as_ref(&self) -> &[u8] {
-        match &self.0 {
-            DataRepr::Binary(d) => d,
-            DataRepr::Text(d) => d
+    /// The length of data (number of bytes).
+    pub fn len(&self) -> usize {
+        match self {
+            Data::Text(n) => *n,
+            Data::Binary(n) => *n
         }
     }
 }
 
-impl AsMut<[u8]> for Data {
-    fn as_mut(&mut self) -> &mut [u8] {
-        match &mut self.0 {
-            DataRepr::Binary(d) => d,
-            DataRepr::Text(d) => d
-        }
-    }
-}
-
-/// Wrapper which restricts the length of its byte slice to 125 bytes.
+/// Wrapper type which restricts the length of its byte slice to 125 bytes.
 #[derive(Debug)]
 pub struct ByteSlice125<'a>(&'a [u8]);
 
-static_assertions::const_assert_eq!(125, crate::base::MAX_CTRL_BODY_SIZE);
-
 /// Error, if converting to [`ByteSlice125`] fails.
-#[derive(Clone, Debug, thiserror::Error)]
-#[error("Slice larger than 125 bytes")]
+#[derive(Clone, Debug)]
 pub struct SliceTooLarge(());
+
+impl fmt::Display for SliceTooLarge {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        f.write_str("Slice larger than 125 bytes")
+    }
+}
+
+impl std::error::Error for SliceTooLarge {}
 
 impl<'a> TryFrom<&'a [u8]> for ByteSlice125<'a> {
     type Error = SliceTooLarge;